Add support for Cache-Control: immutable with a new "immutable" option
More browsers are starting to support Cache-Control: immutable (see: https://hacks.mozilla.org/2017/01/using-immutable-caching-to-speed-up-the-web/). This commit adds support for an "immutable" option to add this directive to the Cache-Control header.

Test Plan: Added unit tests to verify that the "immutable" directive is added only when the option is specified and when the response is successful.

closes #75